Bound the lengths and offsets a directory listing takes from the server
Fixes GO-2026-5051: an out-of-bounds read and panic in ReadDir and
ReadDirPlus. A directory listing against a hostile or man-in-the-middle
SMB server crashes the calling process.

Two mechanisms, both of them a server-supplied number used without being
bounded first.

## The validation overflowed

    len(c) < int(64+c.FileNameLength())

FileNameLength is a uint32 and 64 is untyped, so the addition happens in
uint32 and wraps before the widening. A FileNameLength of 0xFFFFFFFF
makes the sum 63, the comparison passes, and the decoder goes on to
slice the buffer by a length this function had just failed to reject.

On a 64-bit platform the window is 0xFFFFFFC0 upwards, where the uint32
sum itself wraps. On a 32-bit one it is much wider, because int is 32
bits there and the conversion overflows too.

Four decoders had it, all reachable from a server response:
FileDirectoryInformation, FileNameInformation, FileQuotaInformation and
SrvRequestResumeKeyResponse. All four now widen before adding.

## Two offsets were unbounded

readdir advances through entries by NextEntryOffset and reslices by it
without checking it against what is left. conn's compound-response loop
does the same with NextCommand, where an offset inside the 64-byte header
would also point a "next packet" at part of the current one.

Both now stop rather than panic, and both treat a malformed response as
the end of what can be read rather than as a failure: the entries already
decoded are good, and a server that cannot count is no reason to discard
them.

## Provenance

The fix is Arash Payan's, from cloudsoda/go-smb2 commit 7b96c35, which is
the fork the advisory names as fixed. Both repositories are BSD-2-Clause.

Only the security-relevant hunks are taken. The same commit also
gofumpt's several files and adds a panic recovery to a receiver loop that
does not exist in this tree; neither belongs in a fix that wants to be
easy to review.

## Tests

The overflow half is red-green tested with crafted buffers and needs no
server: three decoders accepted a length larger than their buffer before
the change and reject it after, and a well-formed entry — plus one a
single byte short — is checked so the bound rejects what does not fit
rather than everything.

The two offset bounds are applied from the upstream fix but not covered
here, because reaching them needs a server that sends a malformed
response, which this repository has no way to stand up. That is stated
rather than papered over.

+++ internal/smb2/fscc.go
@@ -141,7 +141,7 @@ func (c SymbolicLinkReparseDataBufferDecoder) PrintName() string {
 type SrvRequestResumeKeyResponseDecoder []byte
 
 func (c SrvRequestResumeKeyResponseDecoder) IsInvalid() bool {
-	if len(c) < int(28+c.ContextLength()) {
+	if uint64(len(c)) < 28+uint64(c.ContextLength()) {
 		return true
 	}
 	return false
@@ -302,8 +302,19 @@ const (
 
 type FileDirectoryInformationDecoder []byte
 
+// Widened to uint64 before the addition, not after.
+//
+// The length is supplied by the server. Adding a constant to it in
+// uint32 wraps: a FileNameLength of 0xFFFFFFFF made 64+len come to 63,
+// the comparison passed, and the decoder went on to slice the buffer by
+// a length this function had just failed to reject — an out-of-bounds
+// read, from a directory listing, against any server willing to send it.
+//
+// On a 64-bit platform the window is 0xFFFFFFC0 upwards, where the
+// uint32 sum itself wraps. On a 32-bit one it is far wider, because int
+// is 32 bits there and the conversion overflows too.
 func (c FileDirectoryInformationDecoder) IsInvalid() bool {
-	return len(c) < int(64+c.FileNameLength())
+	return uint64(len(c)) < 64+uint64(c.FileNameLength())
 }
 
 func (c FileDirectoryInformationDecoder) NextEntryOffset() uint32 {
@@ -439,7 +450,7 @@ func (c FileFsFullSizeInformationDecoder) BytesPerSector() uint32 {
 type FileQuotaInformationDecoder []byte
 
 func (c FileQuotaInformationDecoder) IsInvalid() bool {
-	return len(c) < int(40+c.SidLength())
+	return uint64(len(c)) < 40+uint64(c.SidLength())
 }
 
 func (c FileQuotaInformationDecoder) NextEntryOffset() uint32 {
@@ -681,7 +692,7 @@ func (c FileNameInformationDecoder) IsInvalid() bool {
 		return true
 	}
 
-	if len(c) < int(4+c.FileNameLength()) {
+	if uint64(len(c)) < 4+uint64(c.FileNameLength()) {
 		return true
 	}
